Fix for https://issues.apache.org/activemq/browse/AMQ-2391
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@814627 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-core/src/main/java/org/apache/activemq/transport/failover/FailoverTransport.java
@@ -79,7 +79,7 @@ public class FailoverTransport implements CompositeTransport {
 
     private long initialReconnectDelay = 10;
     private long maxReconnectDelay = 1000 * 30;
-    private long backOffMultiplier = 2;
+    private double backOffMultiplier = 2d;
     private long timeout = -1;
     private boolean useExponentialBackOff = true;
     private boolean randomize = true;
@@ -293,11 +293,11 @@ public class FailoverTransport implements CompositeTransport {
         this.reconnectDelay = reconnectDelay;
     }
 
-    public long getReconnectDelayExponent() {
+    public double getReconnectDelayExponent() {
         return backOffMultiplier;
     }
 
-    public void setReconnectDelayExponent(long reconnectDelayExponent) {
+    public void setReconnectDelayExponent(double reconnectDelayExponent) {
         this.backOffMultiplier = reconnectDelayExponent;
     }
